refactored simulate_rf to a separate file and renamed it to _time_domain
