+++ docs/source/_autosummary/zea.rst
@@ -55,6 +55,7 @@ zea
    probes
    parameters
    simulator
+   simulator_time_domain
    tools
    tracking
    utils
