+++ zea/ops/ultrasound.py
@@ -32,13 +32,14 @@ from zea.internal.core import (
 from zea.internal.registry import ops_registry
 from zea.internal.utils import deprecated
 from zea.ops.base import Filter, Operation
-from zea.simulator import simulate_rf, simulate_rf_fast
+from zea.simulator import simulate_rf
+from zea.simulator_time_domain import simulate_rf_td
 from zea.utils import canonicalize_axis
 
 simulator_settings = {
     "exact": partial(simulate_rf, factored=False),
     "factored": partial(simulate_rf, factored=True),
-    "fast": simulate_rf_fast,
+    "fast": simulate_rf_td,
 }
